Usage of async module with curl for making multiple cUrl calls to a web service
Due to Node.js's non-blocking mechanism, making multiple calls to a web
service using cUrl can be restrictive. The most appropriate way to do
this is to make use of async's map function and cUrl module. curl-node
module does not allow multiple sessions. trying to do this in a loop
will result in blocking, and an will cause curl to report being busy.